// src/api_client/models/WriteSetPayload.ts
/* tslint:disable */
/* eslint-disable */
// @ts-nocheck
/**
 * Diem Dev API Specification
 * Diem Dev API is REST API for client applications to interact the Diem blockchain.
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import {exists, mapValues} from "../runtime";
import {
  WriteSet,
  WriteSetFromJSON,
  WriteSetFromJSONTyped,
  WriteSetToJSON,
} from "./WriteSet";

/**
 *
 * @export
 * @interface WriteSetPayload
 */
export interface WriteSetPayload {
  /**
   *
   * @type {string}
   * @memberof WriteSetPayload
   */
  type: string;
  /**
   *
   * @type {WriteSet}
   * @memberof WriteSetPayload
   */
  writeSet: WriteSet;
}

export function WriteSetPayloadFromJSON(json: any): WriteSetPayload {
  return WriteSetPayloadFromJSONTyped(json, false);
}

export function WriteSetPayloadFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): WriteSetPayload {
  if (json === undefined || json === null) {
    return json;
  }
  return {
    type: json["type"],
    writeSet: WriteSetFromJSON(json["write_set"]),
  };
}

export function WriteSetPayloadToJSON(value?: WriteSetPayload | null): any {
  if (value === undefined) {
    return undefined;
  }
  if (value === null) {
    return null;
  }
  return {
    type: value.type,
    write_set: WriteSetToJSON(value.writeSet),
  };
}